Don't call Double.doubleValue(), just let the compiler unbox.
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@9392 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ src/VASSAL/preferences/DoublePreference.java
@@ -49,7 +49,7 @@ public class DoublePreference extends BasicPreference {
     if (value instanceof String) {
       value = Double.valueOf((String) value);
     }
-    defaultValue = ((Double) value).doubleValue();
+    defaultValue = (Double) value;
 
   }
